Added a user trips page and created card view for the data. started to create the trip overview page. added static data for testing. need to fix login so that it passes the current user.

+++ CapstoneTravelApp/CapstoneTravelApp/DatabaseTables/Dining_Table.cs
@@ -19,7 +19,7 @@ namespace CapstoneTravelApp.DatabaseTables
         public DateTime ResDate { get; set; }
 
         [MaxLength(10)]
-        public int ResPhone { get; set; }
+        public long ResPhone { get; set; }
 
         
         public string ResNotes { get; set; }
